+++ D3dxSkinManager.Client/src/modules/workflow/components/index.ts
@@ -2,7 +2,5 @@
  * Workflow components barrel export
  */
 
-export * from './ModImportWorkflowScreen';
-export * from './ModImportQueueScreen';
-export * from './FolderImportButton';
-export * from './WorkflowQueueTable';
+// ModImport components
+export * from './modImport';