+++ D3dxSkinManager.Client/src/modules/workflow/hooks/index.ts
@@ -2,4 +2,5 @@
  * Workflow hooks barrel export
  */
 
-export * from './useModImportWorkflow';
+// ModImport hooks
+export * from './modImport';