+++ D3dxSkinManager/Modules/Workflow/Repositories/IWorkflowRepository.cs
@@ -14,4 +14,6 @@ public interface IWorkflowRepository
     Task<List<WorkflowInfo>> GetActiveByTypeAsync(string type);
     Task UpdateAsync(WorkflowInfo workflow);
     Task DeleteAsync(string id);
+    Task<int> DeleteBatchAsync(IEnumerable<string> ids);
+    Task<List<WorkflowInfo>> GetByIdsAsync(IEnumerable<string> ids);
 }